refactor(security): share bun audit parsing utilities
